fix bug and errors on map view and calendar page

+++ lib/pages/creation/guestnumber_price_page.dart
@@ -57,7 +57,7 @@ class _GuestNumberState extends State<GuestNumber> {
           ),
         ),
         onTap: () => setState(() {
-              prix = data['id'];
+              prix = (data['id'] as int).toDouble();
             }));
   }
